+++ PyHotKey/_platform_stuff/win32.py
@@ -17,6 +17,7 @@
 from pynput._util.win32 import KeyTranslator
 from pynput.keyboard._win32 import Key, KeyCode, Controller, Listener
 
+filter_name = 'win32_event_filter'
 translator = KeyTranslator()
 
 
@@ -28,3 +29,19 @@ def pre_process_key(key):
         if new_key.char or new_key.vk:
             return new_key
     return key
+
+
+def event_filter(self, msg, data):
+    vk = data.vkCode
+    is_utf16 = msg & self._listener._UTF16_FLAG
+    if is_utf16:
+        msg = msg ^ self._listener._UTF16_FLAG
+        scan = vk
+        key = KeyCode.from_char(chr(scan))
+    else:
+        key = self._listener._event_to_key(msg, vk)
+    if msg in self._listener._PRESS_MESSAGES and self._on_press(key):
+        self._listener.suppress_event()
+    elif msg in self._listener._RELEASE_MESSAGES and self._on_release(key):
+        self._listener.suppress_event()
+    return False